Let `LowerLinalgMicrokernels` assume that dynamic dims are contiguous. (#11641)
This is part 2/9 of #11755: Finish the transition to
`--iree-flow-enable-data-tiling`.

In the VMVX backend, when deciding whether an op can be rewritten as a
microkernel call, we need to check if memref layouts inner dimeensions
are contiguous row-major as expected by the microkernel.

When a dimension has dynamic size, that is hard to do as a static
analysis. This PR makes us just assume that dynamic dims are contiguous,
which is safe-ish for e2e workloads because bufferization only creates
contiguous buffers, but is not 100% satisfactory, technically this pass
in isolation is not correct on all possible inputs.

Issue #11633 tracks finding a better solution.