Compute workgroup count based on pack op input shape
The workgroup count computation should not care about the inner
(register-level) tile sizes. But the result shape outer dims of a pack
op depend on tile sizes (they are the quotient of the input shape by
tile shape). With the newly added possibility of dynamic tile sizes,
this code had gotten complicated as the workgroup count computation had
to call the `query_tile_sizes` ukernel. Finally the last straw was that
we realized that since that is host-side not device-side work, that was
the one place introducing a host-side dependency on the VMVX module
providing the `query_tile_size` ukernel, so that was why we suddenly
needed either #11643 or #11644.