[stream] Don't group timepoint awaits in different blocks
We sort various await ops over the same timepoint according to
their order in the same block, but the list was pulling in await
ops in other blocks too. This caused crash.

In general, unifying ops across the blocks is triky given we need
to consider dominance and referened values in each op. Prefer to
be conservative for now.