Commonolize implementation of transform interpreters (#11810)
Factor out the common components of TransformDialectInterpreterPass in
Codegen, in iree-dialects, and of DispatchWithTransformDialect into a
base class template that lives in iree-dialects. This includes the logic
necessary for injecting transforms from a file, maintaining those alive
in a separate module, debug targeting via attributes, and transform
module identification in the IR. It is currently used for IREE codegen,
and will be reused for Flow and standalone interpreter.