Add ireeCompilerInvocationSetCrashHandler compiler API.
This uses the callback based mechanism for generating crash reproducers and exposes it through the C API in a way that I expect to be able to use easily from the PJRT plugin.

Testing this is beyond the basic test setup that we have for the API, and we should really aim to use this from the iree-run-mlir replacement (which will be a proper JIT based on the API) in order to exercise these corners properly.