[tracy] Add support for `csvexport` tool.
This tool provides information about the trace in machine-readable form.
This tool loses a lot of context from the trace, so it is recommended to
only use this tool in reference to a particular trace that has been
already examined manually or otherwise controlled so that the extracted
numbers are known to be meaningful.