Support transform dialect matchers in dispatch region formation
This requires some changes to the base logic of the transform dialect
interpreter pass. The transform IR is constructed on the fly in a
separate module, the ownership of which is shared by all copies of the
dispatch-with-interpreter pass, similarly to the externally supplied
transform IR file. The interpreter currently is currently run multiple
times with Linalg ops as potential roots to control the matching from
C++ instead of deferring it to the transform dialect.