[benchmarks] Add a host CPU for x86 benchmarks
Prior to this patch, the binaries produced for the x86 benchmarks would run
only on `CascadeLake` capable hardware.
This may be annoying when running experimentations on older/different x86
hardware.

To make the experience smoother in that space, add a `host` target.

Aside from the addition of the `host` target, this patch would be NFC except
for one thing:
- Now all the TFLite configurations run `MobileNetV1`.
Previously only one did. I assume it wasn't intentional.