Rename away from `AArch`-everything (#11903)
A number of things in LLVMCPU/ unnecessarily have references to AArch64
in their names.

It's most obvious in the `setRootConfig` overload taking a `Mmt4dOp`:
this function is called regardless of architecture for any `Mmt4dOp`,
and yet it ends up selecting `CPUAArchDoubleTilingExpert`:
https://github.com/iree-org/iree/blob/a9227685614999b4de7ee53b3843ef1c7b409489/compiler/src/iree/compiler/Codegen/LLVMCPU/KernelDispatch.cpp#L993-L1038

So the chain of renamings in this PR starts here. There are multiple
problems in the name `CPUAArchDoubleTilingExpert`:
1. It's suggesting something specific to ARM architectures ("AArch")
even though it's used everywhere.
2. Just `AArch` alone without a 64 or 32 suffix is not a common term.
3. The `Double` part is out of sync with what the code actually does (it
uses a `SingleTilingExpert`).

So, in line with other experts in the same file, we rename
`CPUAArchDoubleTilingExpert` to `Mmt4dTilingExpert`. This is consistent
with the above described call site; there is only one other place that
creates that expert, that was `setAArch64RootConfig` in the same file.
That one was actually specific to `AArch64` but not Mmt4d, so the new
name looks out of place here, but I think that's OK because this should
only be used in non-data-tiling on AArch64 and I think we already know
that this is going away soon.

The rest of this PR follows by similar renamings to be consistent.

This results in over-using the term `Mmt4d` a bit, since for instance
the vector lowerings only see `vector` ops such as the `vector.contract`
that result from vectorization of `mmt4d`, not the `mmt4d` themselves.
But I think that's less bad than overusing `AArch64` in code that runs
on all CPU architectures.

Very much open to better names --- don't want to blindly jump from
aarch-everything to mmt4d-everything. Maybe a follow-up would reduce the
number of separate experts, and the new name would arise for the more
unified expert?