Adding debug information to LLVM CPU codegen outputs. This restores line table information that had been dropped due to upstream MLIR changes by adding DISubprogramAttr to dispatch functions. In addition the arguments commonly used within dispatch functions are annotated with value information for use by debuggers. Additional debug information can be added in the future but hopefully with a way to more automatically keep it in sync with the runtime definitions. Since all of the debug information is strictly informational it's ok for them to get desynchronized (similar to our natvis file).
Internal functions are not yet annotated but we don't generate any with
the current pipeline anyway. Externs/imports are also not annotated but
they are not usually needed by debuggers as the destination will usually
contain this information or it'll be brought in via bitcode we merge.

Since this added quite a bit more code the debug info and type/accessors
have been moved to their own file and cleaned up a bit. Previously we
would recreate and cache the ABI accessors for every op converted leading
to some expensive attribute interning - now we only cache the debug info
and types once and reuse that for all ops.