Adding a greedy packing transform dialect op.
This PR adds a transform.iree.pack_greedily operation that infers the packing for 2-D contractions
embedded in any contraction op and packs accordingly.

A normalization step guarantees that we get the innermost op dimensions in one of 8 possible
`(m, n, k)` orders, specified as a parameter, from which we can emit all packed forms.

The current implementation takes an arbitrary linalg op and tries to pack it along dimensions
ordered as `kk, mm, nn` and with sizes `32, 8, 16`.
This is an arbitrary choice that can be lifted to a transform.pack_greedily.

This achieves a new level of normalization and generalization for any n-D linalg op that contains
a contraction hidden within it: we will always see a predictible packed form for any of these ops.

Followups will generalize to convolutions, reductions and pointwise.