Adding a greedy packing transform dialect op.
This PR adds a transform.iree.pack_greedily operation that exercises the mechanics
of packing for 2-D contractions.

A normalization step guarantees that we get the innermost op dimensions in the `(m, n, k)` order
after which we simply brute force the packing.

The current implementation takes an arbitrary packing size of `8x16x32` and an arbitrary ordering
of outer and inner dimensions.

Followups will generalize to n-D contractions, convolutions and reductions.