Adding CUDA device-side tracing support (`--cuda_tracing`). (#11943)
This uses a pool of CUevents and trace zones inserted during stream
execution and is currently capped at a reasonable amount for most models
but it's something we can make dynamic in the future if needed. All
zones are collected every synchronize which in today's synchronous CUDA
driver is after every submission completes. In the future we'll want to
more lazily process the events and only optimistically poll on
submissions however since the time it takes to get the event times is so
high (~10us/event) we'll probably want to build out a dedicated polling
thread so that timing is not too skewed. Today several milliseconds per
submission will be spent collecting timings and total execution time
numbers will be inflated and users must opt-out if they want traces with
accurate host timings (we could change this if we find it confusing).
Additionally some of the more expensive tracing work has been moved
behind the `IREE_HAL_VERBOSE_TRACING_ENABLE` flag that by default is
only on in builds with assertions enabled as the user is indicating they
don't need precise host timings.

Graphs will require a bit more plumbing to allow for updating the
events/trace zones used during launches but since most performance work
is being done with streams this is deferred. The idea is to record event
nodes in the graph during command buffer recording and build a table
retained on the command buffer of CUgraphNodes and the trace data - then
each time a graph is launched we'd walk the table and record new events
then use cuGraphExecEventRecordNodeSetEvent to update the nodes before
launching. In this way we can reuse the graphs with unique timings per
launch.

Example with annotations:

![image](https://user-images.githubusercontent.com/75337/214175998-dbcaf992-62af-4105-8409-ea178d840a28.png)
* GPU execution time: total time of command buffer (beginning of first
command to end of last)
* GPU self time: all time within the total GPU execution time not
attributed to dispatches (copies/fills/overhead)
* CPU command setup time: total time on host issuing CUDA API calls
(beginning of first command to end of last)
* Delay to execution: **ESTIMATED** time between the first command host
API call and the start of its execution on device
Note the non-trivial tracing overhead at the end of the submission!

Clicking on the command buffer zone will show all dispatches sorted by
GPU time descending. Hovering over any in the list will show where they
are on the timeline in the main display. It's useful to check "Group
children locations" when evaluating dispatch performance though note
that the same kernel may be dispatched with different grid sizes!

![image](https://user-images.githubusercontent.com/75337/214176725-2060a772-3bbe-4764-97dd-0257b1a56d01.png)

Clicking into any particular dispatch will highlight the host call that
created it - in the future we can attach more information to these
events at the cost of additional overhead when tracing (maybe a verbose
tracing mode or something) - for now the workgroup count is only
attached in builds with assertions enabled or
`IREE_HAL_VERBOSE_TRACING_ENABLE` forced due to the expense:

![image](https://user-images.githubusercontent.com/75337/214177870-efb55976-1fbb-4d01-9668-b2b66fea1413.png)