Make stack allocations checks in generated code on LLVMCPU more stringent (#11938)
Earlier checks were ensuring that the stack allocation sizes are bounded, but without enforcing that these are hoisted all the way to the top this could still lead to explosion of stack allocations. Add this check. This gives LLVM codegeneration a static stack frame. With that is guaranteed to be bounded.

Issue #11880