Add an example of generalized packing
This revision demonstrates how the generalized packing transformation is a one-size fits all implementation
for tiling linalg ops of rank N to linalg ops of higher rank.

The tensor.pack / tensor.unpack representation allows us to complete our panoply of composable linalg transformations.
Now, one can either:
1. lower a linalg op to loops and use classical loop-based tiling techniques (e.g. Allen&Kennedy, polyhdral/affine etc)
2. tile an N-D linalg op to N-D loops surrounding an N-D linalg op. This often preserves the name of the linalg op and is at the basis of the TilingInterface.
3. (this PR) tile an N-D linalg op to a 2*N-D linalg.generic. This step requires that the tile dimensions divide the problem dimension. tensor.pack / tensor.unpack provide this guarantee.

Step 3. can easily be adapted to produce a new named op (e.g. mmt4d) when relevant, the point of this PR is to demonstrate generality.

This is related to discussion #12075.

An additional pattern is added to convert tensor.pack/unpack to linalg_ext.pack/unpack until IREE adopts the upstream variants.
With this, it is possible to form dispatch regions without failing to lower.

One thing to note in the IREE pass pipeline is that the InterchangeGenericOps breaks the normalization property of the packing.
This can be recovered after the fact but it would be better to disable in such cases if possible.

At this time, `iree-compile` fails on the `iree_linalg_ext` ops as it wants statically allocated buffers to be hoisted to the top of the function.

Both these issues can be left for a followup investigation.