fix bad stride in pack_benchmark (#12009)
Fixes #12008
Regression introduced in #11987

I had tweaked the strides in the benchmark as I thought for a moment
that it was more realistic to benchmark round (power of two) strides
even when padding makes sizes a little smaller. But now I'm not so sure
anymore what's the more important case in practice, so this just reverts
to the old state.

The crash was because instead of the correct ` out_size1 * out_size3`
(value for `in_size1` before adjustment for padding) I had incorrectly
typed `out_size0 * out_size2`.