Replace 'zve32x' with 'zve32f' in CI configurations
The 'zve32x' feature includes support for integer vectors but not fp
vectors. When we send fp vector ops to the RISC-V backend with only
'zve32x' support, they have to be scalarized. This is currently exposing
some implementation gaps in the backend which are hopefully addressed
soon (see https://github.com/llvm/llvm-project/issues/60463). In the
meantime, we are replacing 'zve32x' with 'zve32f', which also includes
support for fp vector operations.