Transform: avoid rematching ops when already captured (#12072)
When running the transform dialect-related matcher recursively, check if
a corresponding payload operation has been already captured. If so,
check whether the inspected operation is the same as already captured
instead of rechecking the predicates. This allows matchers to be used
repeatedly in a recursive structure to indicate that the same operation
must be used in several places.