[wip] hook up ptx assembly injection
This is a quick-and-dirty proof of concept of how we could enable bottom-up
investigations.
For this patch, we hook up only the lowest level available for our cuda GPU.

The idea is to be able to easily prove that something works before
implementing the right support in the compiler.

**Context**

When looking at performance issue (or correctness!), it may be costly to
actually teach the compiler to do the right thing before we actually know
if it is worth it (or fixes the issue!). Therefore we want to provide entry
latches to check if the change we have in mind is really worth implementing or
to see how to get the desire output at a level closer the end result.

The current patch is limited, obviously, and for instance some changes may not
be possible at this level, but again, proof of concept + it can be useful
already!

Long term, the plan is to create entry points at every level of compiler to
allow prototyping/investigating at the right level of abstraction.

This is not something new, for people familiar with the low level LLVM
compiler (`llc`), this is how `start-before` and co work. For people not
familiar with `llc`, essentially these options allow to start/stop the
compiler pipeline anywhere so you can reproduce issues, check
hypothesis, experiment, etc in a setting that is close to your actual
compiler. (This is actually your compiler, you're just starting/finishing in a
different position in the pipeline.)

Note: I won't go into the details of how this could be implemented properly in
general, because I don't know how similar to LLVM is the pass manager in MLIR.

**Current Implementation**

The way it works is pretty dumb:
In a nutshell, we tell the compiler to replace different kernels with the
content of different files. The content of the files is expected to be PTX
code.

We provide this information via a compiler option,
`iree-hal-cuda-read-ptx-from-file=<pathToMappingFile>`, that points to a
text file with the desired mapping:
```
<functionName1>,<pathToPTXReplacement1>
...
<functionNameN>,<pathToPTXReplacementN>
```

Note: Based on past experience, the mapping file indirection here is useful
(instead of passing all that in an option for instance), because it is easier
to comment out/add entries in the file (the format supports full line comment
with `#`).

Note: We could probably achieve the same effect than what this patch does
by patching up the object/external attribute on `ExecutableExportOp`. In
this patch I do something dumber that doesn't require any IR update/patching.
The rationale for this (prototype) choice is that injecting IR shouldn't
require to change the compiler passes, only the pass manager should be impacted
and as result I didn't dig into patching the IR.

**How to Use It?**

Note: If you are a PTX hardcore, you can skip the first two items and
instead write/generate with other tools your ptx files directly!

* First generate the base PTX input for your program:
```
iree-compile --iree-hal-target-backends=cuda <...> --iree-hal-cuda-dump-ptx > basePTX.ptx 2>&1
```

* Extract the functions you want to modify in their own files, e.g.,
`kernel1.ptx` for your first function, `kernel2.ptx` for the second one and
so on.

* Modify your `kernelX.ptx` all you want, but beware that you must keep a
  stable API with the rest of the world. I.e., what's visible to the outside
  world must remain the same.

* Create a mapping file that lists all the kernels you want to replace and
  where to find their code. The content of the mapping file should look
  like:
```
myFunc1,path/to/kernel1.ptx
myFunc2,another/path/for/kernel2.ptx
```

* Rerun `iree-compile` with the same options as the original invocation but
ditch `--iree-hal-cuda-dump-ptx` and add
`--iree-hal-cuda-read-ptx-from-file=mappingFile.txt`.

Currently if any replacement occurs, you should see
`Found: 'kernel.ptx' for 'funcName'` and no error message.

Note: You can keep `--iree-hal-cuda-dump-ptx` if you want to check that
your hand coded PTX files have been properly injected.

* Finally run your "hand modified" code with `iree-run-module` like you
  would normally do.

**Limitations**

* We assume one function per kernel
* When we match a function we replace the whole kernel (i.e., if the kernel
  has several functions, we'll replace all of them with the kernel file of
  the first function that we found in the mapping file.)
* We do the matching on the name of the IREE dispatch functions, not the LLVM
  uniqued symbols. In general these are the same as the final PTX file, but in
  some instances they could change.
* We currently parse the mapping file again and again for each dispatch.
  This could be a compile time sink, but in practice, we should use that
  only for prototyping.