Perform GPU address space lowering during LLVM lowering
This patch removes the standalone GPU address space lowering pass and
performs that conversion from the address space enum during the
respective LLVM conversion passes. That matches upstream LLVM. This
allows the removal of a hard-coded address space constant from the
conversion of shared memory AllocOps to global variables. The correct
address space has to be provided as an option for the pattern.