Add a matcher for value packs
It is sometimes useful to define matcher predicates on a pack of values
without knowing the number of values in the pack. For example, when
all results of an operation are used by another operation, regardless of
their number.

This required a small refactoring of the base matcher code to avoid base
matcher knowing about all derived classes and using complicated logic
for resetting the captured values.

Use this to replace the custom hook for a structured op being a
pass-through, i.e., for its body yielding the input block arguments.