Update `ResolveBufferDescriptors` to handle `memref.extract_strided_metadata`
The pass currently is meant to handle
vmvx.get_buffer_descriptors. This operation is very similar to the
memref.extract_strided_memref and the logic could be re-used to handle
this operation as well. This allows using the ukernel path being added
to IREE which is intended to work for both VMVX and LLVM CPU
codegeneration paths.