Update `ResolveBufferDescriptors` to handle `memref.extract_strided_metadata` (#12205)
The pass currently is meant to handle `vmvx.get_buffer_descriptors`. This operation is very similar to the `memref.extract_strided_memref` and the logic could be re-used to handle this operation as well. 
This allows using the ukernel path being added to IREE which is intended to work for both VMVX
and LLVM CPU codegeneration paths.

The only pattern that is a bit more harder to reuse is

```
%0 = hal.interface.binding.subspan
.. = vmvx.get_buffer_descriptor %0
```

since the base buffer used by the `vmvx.get_buffer_descriptor` isn't the same type as `memref.extract_strided_memref`. That pattern is not adapted yet.