Separate the compiler stable API into compiler/bindings/c.
Prior to this, our public headers and loader were in the internal implementation directory. Aside from being layered wrong, this creates practical problems because the headers are only available when building the compiler (but are usable without a locally built compiler via dynamic linking).

This patch moves the embedding and tool entry point APIs to compiler/bindings/c and declares them stable at their new include paths:

* `iree/compiler/embedding_api.h`
* `iree/compiler/loader.h`
* `iree/compiler/tool_entry_points_api.h`

These can be referenced with the new targets `iree::compiler::bindings::c::headers` and `iree::compiler::bindings::c::loader`. Both are available regardless of whether the compiler is built.

We leave the `MLIRInterop` and MLIR-C API as exported but private (and unstable) from the perspective of IREE's public API.

Progress on #12326