[LLVMGPU] Enable tensor.pack op e2e execution on cuda.
It takes 2.7 ms to finish the tensor.pack e2e tests. This is because the
fully dynamic tensor.pack test cases are not vectorized. There are two
fully dynamic cases. If we exclude the test, it takes ~0.57 ms. The
pipeline is not well tuned yet, it is mostly copied from LLVMCPU and for
correctness tests.