Introduce a `iree_codegen.get_base_pointer` operation.
This operation allows getting the base pointer of a `memref`. This is
similar to `memref.extract_strided_metadata` but extracts just the
base pointer. This is useful for handling `alloca`s and `global`s
during lowering to micro kernel that requires the base pointer of
these objects. Previously a `builtin.unrealized_conversion_cast` was
used to implement this behavior