Use Release mode for GCC CI Build (#12398)
The GCC build directory is getting huge and we never run tests with it
(since this is an ancillary build to ensure this compiler configuration
works). It's gotten so big that we're running out of disk (really tmpfs)
space on the CI runners. From my local testing, this cuts the disk usage
for the build directory after build_all.sh from ~100GB to ~4GB.