support unspecified channel_handle and use_global_device_ids
When channel_handle and use_global_device_ids are not specified,
they default values are 0 and false, respectively.

When they are unset, the communication type is `cross-replica`.
In particular, when there is a single replica group, the cross-replica
is the same as the flatten_ids, so the case can be supported now.