Rework `cpu.c` to make room for x86 path (#12395)
The main change is to switch the nesting of the `#if`'s. This used to be
first `#if` on platform/OS, then `#if` on CPU architecture. But actually
the fact that the platform/OS needs to be involved is a ARM-ism; it is
the CPU architecture that drives the higher level approach. And
structuring the code this way makes it possible to later split it into
separate files for each architecture (not doing it for now because this
opens questions of what we do with the Bazel+CMake build system, so
deferring that until there's a concrete need).

And even for the existing code, having the `#if` on CPU architecture at
the top results in simplification as a couple of platform/OS-centric
functions don't need a fallback anymore now that they only exist within
a CPU-architecture-specific section.

Also flattened some helpers, IMO making code more readable.

Also renamed cpu_data identifiers to be shorter.