[Common][TransformDialect] Fix the usage of num_threads for TileToForallAndWorkgroupCountRegionOp
Prior to this patch, it was impossible to use
`TileToForallAndWorkgroupCountRegionOp` without specifying the tile sizes. The
lowering of the workgroup_count op would be simply skipped leading to compiler
errors down the line.

This patch fixes that by teaching `lowerWorkgroupCountComputingRegion` (the
function that lowers workgroup_count ops) how to get the workgroup sizes from
`num_threads`.