Adding compilation reentrancy tests and new HAL pipeline phases.
Two new `--compile-to=` phases are supported:
- `executable-sources`: run just past interface materialization where
  `hal.executable` ops with target configurations are present.
- `executable-targets`: run just past executable translation where
  `hal.executable.variant` ops have been lowered to their final MLIR form
  before linking and serialization (LLVM dialect, SPIR-V dialect, etc).

New tests are added that demonstrate and verify that iree-opt pipelines
can be run piecewise to produce a final output and that `--compile-to=`
at any particular phase can be passed back in and lowered down to a
final output.

A few passes were tweaked to ensure these tests pass.