Fully retire CanonicalizedSequenceOp
This revision fully retires the CanonicalizedSequenceOp and makes the need for
canonicalization / tiling_canonicalization / licm / singleIterationLoopPromotion
explicit in the transform IR.

This removes automatic blanket pattern applications which is expected to be beneficial for:

- reducing compile time
- reducing surprise (explicit being better than implicit)
- help surface issues similar to Intermittent failures in mobilenetv3_fake_weights.mlir #12444

Add a buildCanonicalizationAndEnablingTransforms and remove CanonicalizedSequenceOp

In the process, plumb the isolatedFromAbove through the API to make the requirement explicit
until we are able to remove it in some future (i.e. with better targeted canonicalization and
greedy pattern rewriter that does not require an isolated from above).