Starting support for HAL dispatch specialization. This allows `stream.cmd.dispatch` ops to specify multiple entry points in executables that can be dispatched. During interface materialization the entry points are expanded for all materialized variants. HAL backends are also able to add their own entry points either in existing variants or new ones during translation (no helpers yet, but it's possible). When lowering the `stream.cmd.dispatch` into `hal.command_buffer.dispatch` each export now chooses the condition for which it should be selected.
This initial work just prepares the op for this conditional dispatch and
also solves an issue with compiler re-entrancy where between interface
materialization and HAL conversion we had IR that would not verify. Only
the existing target variant condition is supported but we now have the
place where we'd call out to HAL backend-produced logic for selection
(getExportConditionAttr).

The `inline-dynamic` HAL model can support this too but for now that's
deferred until we add the backend support - the `hal.device.switch` op
needs a reworking to be used in that context.