Add bazel `iree_link_compiler_shared` flag.
With prior changes, it is now possible to make Bazel mirror CMake with respect to shared linking of the compiler. This flag does that. We default it to false for safety (CMake defaults to true). I don't know how much we will use this but having 1:1 correspondance of the build system has some merit.