Add a pass to correctly represent hal.interface.binding.subspan ops with offsets.
`hal.interface.binding.subspan` ops of the form

```
%subspan = hal.interface.binding.subspan set(0) binding(0) offset(%o0)
    : memref<?x?xf32>{%s0, %s1}
```

are to be interpreted as follows.

- `hal.interface.binding.subspan set(0) binding(0)` is effectively `char *binding`.
- A non-zero offset offset(`%o0`) corresponds to performing `binding + %o0`.
- The bytes of data from `[binding + %o0, binding + %o0 + %s0 * %s1 * <elem-type-byte-width>)` can be viewed as a `memref<?x?xf32>{%s0, %s1}`.

For backends that have a operations which can perform pointer
arithmetic (like gep instructions in LLVM), this interpretation can be
directly lowered. Backends like VMVX end up using hal.buffer.subspan
operations to represent the pointer+offset. This operation does not
get lowered to VM dialect directly, but are rather folded away with
uses. For cases where we need to get a raw pointer + offset value,
these operations dont get folded with uses and fail compilation.

One way to handle this is to convert these operations to represent the
entire subspan under memref semantics. So the above operation can be
represented as

```
%subspan = hal.interface.binding.subspan set(0) binding(0) offset(%o0)
    : memref<?x?xf32, strided<[?, 1], offset : ?>>{%s0, %s1}
```

giving the subspan effectively subview semantics. This allows use of
operations like memref.extract_strided_metadata to resolve the
offsets, sizes, base pointer usages, etc.

To implement this path, two things are done
1) A new pass is added to convert subspan with offsets into the above description.

2) Leaving the subspan as still doesnt solve the problem. The offset
   added can then be folded into the size of the subspan when the
   subspan is linearized.
```
%offset = affine.apply affine_map<()[s0] -> (s0 floordiv 4)>()[%o0]
%size = affine.apply affine_map<()[s0, s1, s2] -> (s0 * s1 + s2 floordiv 4)>()[%s0, %s1, %o0]
%subview_size = affine.apply affine_map<()[s0, s1] -> (s0 * s1)>
%subspan = hal.interface.binding.subspan set(0) binding(0) offset(%c0)
    : memref<?xf32>{%size}
%subview = memref.subview %subspan[%offset] [%subview_size] [1]
    : memref<?xf32>{%size} to memref<?xf32, strided<[1], offset : ?>>
```

This is implemented in `FlattenMemrefSubspanPass` and further cleanups
to plumb this through.

A consequence is that all backends that do not support GEP semantics
now need to use the `FixupSubspanOffsetsPass` and
`FlattenMemrefSubspanPass` in conjunction.

In the long term, all backends should handle offsets to subspans
similarly and drop the use of GEP in the LLVM lowering.

Fixes #12480