Rework CI setup variables (#12524)
`ci-stage` was a confusing concept that conflated a few different
things. I split those things out instead. We've had a few issues
related to this in the last week:
https://github.com/openxla/iree/pull/12482 and
https://github.com/openxla/iree/pull/12514, and posting benchmark
comments is still failing because that's triggered by `workflow_run`:
https://github.com/openxla/iree/actions/runs/4346904097.

This includes making booleans json rather than magic strings.
`write-caches` is still "0" or "1" because it's ultimately used by
bash.