[spirv] Make confugration deduction return state consistent  (#12561)
We have some kernel confugration deduction function (e.g., for
reduction) returning `failure()` on failing to find a configuration, and
the rest returning `success()`. This is inconsistent and confusing.

Additionally, for the latter we rely on `getLoweringConfig()` as a
separate call to check whether a configuration is indeed set. This won't
actually work anymore with transform dialect, where we will embed the
configuration in the schedule.

So this commit switches all into returning `failure()` meaning unable to
find a configuration.

This is a preliminary step towards enabling transform dialect in SPIR-V
CodeGen.