Hack to rewrite address computations using affine.apply
This composes with Nicolas' changes to decompose affine.apply into
`affine.apply`s that are loop dependent and loop independent. This allows to
hoist the loop independent part without relying on LLVM reassociation.

This proof of concept is baked in the gpu pipeliner, but it should be its own
thing. We also probably want something that is not so tied with the llvm
conversion.