[LLVMGPU] Optimize shared memory allocation size
In order to improve performance for matmul with large tile sizes we need
to be able to re-use shared memory between the allocation done for the
input and output.
This adds a pass to pack memory allocations if we can prove that their
liveranges are disjoint. This relies on a simplisitic algorithm that will
sort allocations into alias groups that are independent of each others.
Once we found groups of disjoint allocs we pack them into a single byte
allocations and use memref.view to break it down into several views that
can replace the existing allocations.