[LLVMGPU] Optimize shared memory allocation size (#12538)
In order to improve performance for matmul with large tile sizes we need
to be able to re-use shared memory between the allocation done for the
input and output.
This adds a pass to pack memory allocations if we can prove that their
liveranges are disjoint. This relies on a simplisitic algorithm that
will sort allocations into alias groups that are independent of each
others.
Then we create a single alloc that is partitioned for each independent alias group. This way we only use the amount of memory needed by the largest alias group.